refactor: extract transformers and claim modules from storage.ts
storage.ts: 1063 → 876 lines (-18%)

Extracted:
- transformers.ts (71 lines): rowToRun, rowToStep, rowToLog, validateLabels
  Pure data transformation, zero DB dependency.
- claim-postgres.ts (95 lines): FOR UPDATE SKIP LOCKED + advisory locks
- claim-sqlite.ts (54 lines): atomic UPDATE with subquery

claimNext in storage.ts now delegates to the appropriate backend module.
Each file has a single, clear responsibility.

Closes #131
